use system zlib instead of included package
This would have prevented being vulnerable to CVE-2018-25032 and CVE-2022-37434
but at least makes us immune to future problems with the bundled zlib.

When hrsync is used with the options --chroot and --dropsuper it is believed
exploitation by a remote host could only affect the files from that host in the
given chroot and not any files in another part of the file system.